requirements.txt
Babel==2.5.1
brotlipy==0.7.0
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-appconf==1.0.2
django-autoslug==1.9.3
django-cacheops==4.0.3
django-compressor==2.2
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-fastsitemaps==0.3
django-filter==1.1.0
django-haystack==2.6.1
django-phonenumber-field==1.3.0
django-pure-pagination==0.3.0
django-redis==4.8.0
django-suit==0.2.25
django-suit-redactor==0.0.4
djangorestframework==3.7.7
drf-extensions==0.3.1
drf-haystack==1.6.1
elasticsearch==6.0.0
funcy==1.10
idna==2.6
oauthlib==2.0.6
phonenumbers==8.8.8
phonenumberslite==8.8.8
psycopg2==2.7.3.2
pyinotify==0.9.6
python-dateutil==2.6.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.3
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
whitenoise==3.3.1
Whoosh==2.7.4
